XMLHttpRequest/send-sync-no-response-event-load.htm

<!DOCTYPE html>
<html>
<head>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-send()-method" />
    <script src="/resources/testharness.js"></script>
    <title>XMLHttpRequest L2: The send() method: Fire an event named load (no response entity body and the synchronous flag is set)</title>
</head>

<body>
    <div id="log"></div>

    <script type="text/javascript">
        test(function()
        {
            var xhr = new XMLHttpRequest();
            var pass = false;

            xhr.onload = function(e) 
            {
                assert_true(e instanceof ProgressEvent);
                assert_equals(e.type, "load");
                pass = true;
            };

            xhr.open("POST", "./support/echo.php", false);
            xhr.send();

            assert_equals(xhr.response, "");
            assert_true(pass);
        });
    </script>
</body>
</html>
